Fix validation rendering in forms.
Realised that the serverside form validation errors markup was slightly wrong and that they were being hidden by CSS rules accidentally.

Change adds the is-invalid class to the form input if there are problems and moves the validation errors to be immediately next to the input. With this change they should show properly on the page.